Debug refactor merge
