+++ vscode/src/lsp/requests/register.ts
@@ -22,4 +22,4 @@ export const registerRequestListeners = (client: NbLanguageClient) => {
         const { type, handler } = listener;
         client.onRequest(type, handler);
     })
-} 
+}
